Moved some methods between CraftCommons and Util classes.
