+++ src/main/java/com/craftfire/commons/exceptions/AnalyticsException.java
@@ -19,7 +19,6 @@
  */
 package com.craftfire.commons.exceptions;
 
-import com.craftfire.commons.CraftCommons;
 import com.craftfire.commons.Util;
 import com.craftfire.commons.classes.AnalyticsData;
 import com.craftfire.commons.managers.AnalyticsManager;
@@ -41,6 +40,6 @@ public class AnalyticsException extends Exception {
     }
 
     public int getResponseCode() {
-        return CraftCommons.getResponseCode(getManager().getURL());
+        return Util.getResponseCode(getManager().getURL());
     }
 }
